Add files via upload
Introducing "EasyBank 🏦": a concise C++ console application offering a simplified banking experience! Users log in with their name and a secret password (currently set to 7078). Once inside, they can effortlessly check their balance, deposit funds, withdraw cash, and gracefully exit. The initial balance is set to $100, and the code is designed to be beginner-friendly. It's a perfect starting point for those looking to explore basic banking operations in C++. Get ready to dive into the world of "EasyBank"! 💳✨